More code cleanup: - Collapsed catch blocks that did the exact same things - Replaced ArrayList<Class> with ArrayList<> - Replaced StringBuffer with StringBuilder - Removed a few usages of deprecated classes (mainly HSSFColors in Excel) - Replace Collections.sort with List.sort

+++ projectforge-business/src/main/java/org/projectforge/business/ldap/LdapPosixAccountsUtils.java
@@ -57,7 +57,7 @@ public class LdapPosixAccountsUtils
       if (ldapUserValues == null) {
         continue;
       }
-      if (ldapUserValues.getUidNumber() != null && ldapUserValues.getUidNumber().intValue() > currentMaxNumber) {
+      if (ldapUserValues.getUidNumber() != null && ldapUserValues.getUidNumber() > currentMaxNumber) {
         currentMaxNumber = ldapUserValues.getUidNumber();
       }
     }
@@ -82,7 +82,7 @@ public class LdapPosixAccountsUtils
         continue;
       }
       if (ldapUserValues != null && ldapUserValues.getUidNumber() != null
-          && ldapUserValues.getUidNumber().intValue() == uidNumber) {
+          && ldapUserValues.getUidNumber() == uidNumber) {
         // Number isn't free.
         log.info("The uidNumber (posix account) '" + uidNumber + "' is already occupied by user: " + user);
         return false;
